Added image display for food items
Added an optional image_url value to the food items which hold an image url and this can be displayed on the food item card if present. I have also updated the in stock status text on the cards to indicate out of stock if the checkmark is false. I also slightly edited text on the inventory site. I have also added the image_url value in the add items form, but eventually we will be uploading an image directly instead of just inputting the url.