Updated style of edit button and other changes
Updated style of edit button. Changed instances of class to className. Changed upload image since the previous one took too long to load.
There is a bug with the display of the checkbox status when editing. It seems to show the previous action taken but doesn't show the correct status when clicking edit item. However, the form response seems to work ok.