Almost working feature for uploading images to add item
Currently able to upload images to add item and it uploads the image to firebase and gets the url and adds it to the image url so it properly adds the new food item with the image. Currently it's not formatted and if the user doesn't upload an image, it will attach the previous image uploaded.